Remove classic KubeCluster and HelmCluster
