--- .github/workflows/helmcluster.yaml
@@ -1,57 +1,0 @@
-name: "HelmCluster"
-on:
-  pull_request:
-    paths:
-      - ".github/workflows/helmcluster.yaml"
-      - "requirements*"
-      - "ci/**"
-      - "dask_kubernetes/helm/**"
-      - "dask_kubernetes/common/**"
-      - "dask_kubernetes/*"
-  push:
-    paths:
-      - ".github/workflows/helmcluster.yaml"
-      - "requirements*"
-      - "ci/**"
-      - "dask_kubernetes/helm/**"
-      - "dask_kubernetes/common/**"
-      - "dask_kubernetes/*"
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    timeout-minutes: 15
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
-        kubernetes-version: ["1.29.2"]
-        include:
-          - python-version: '3.10'
-            kubernetes-version: 1.28.7
-          - python-version: '3.10'
-            kubernetes-version: 1.27.11
-          - python-version: '3.10'
-            kubernetes-version: 1.26.14
-
-    env:
-      KUBECONFIG: .pytest-kind/pytest-kind/kubeconfig
-
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install deps
-        run: ./ci/install-deps.sh
-      - name: Run tests
-        env:
-          KUBERNETES_VERSION: ${{ matrix.kubernetes-version }}
-        run: pytest --reruns=5 dask_kubernetes/common/tests dask_kubernetes/helm/tests
-      - name: Debug k8s resources
-        if: success() || failure()
-        run: kubectl get all -A
