--- dask_kubernetes/classic/__init__.py
@@ -1,2 +1,0 @@
-from ..common.objects import make_pod_spec
-from .kubecluster import KubeCluster
