--- dask_kubernetes/classic/tests/config-demo.yaml
@@ -1,38 +1,0 @@
-apiVersion: v1
-clusters:
-- cluster:
-    certificate-authority: fake-ca-file
-    server: https://1.2.3.4
-  name: development
-- cluster:
-    insecure-skip-tls-verify: true
-    server: https://5.6.7.8
-  name: scratch
-contexts:
-- context:
-    cluster: development
-    namespace: frontend
-    user: developer
-  name: dev-frontend
-- context:
-    cluster: development
-    namespace: storage
-    user: developer
-  name: dev-storage
-- context:
-    cluster: scratch
-    namespace: default
-    user: experimenter
-  name: exp-scratch
-current-context: dev-frontend
-kind: Config
-preferences: {}
-users:
-- name: developer
-  user:
-    client-certificate: fake-cert-file
-    client-key: fake-key-file
-- name: experimenter
-  user:
-    password: some-password
-    username: exp
