--- dask_kubernetes/helm/__init__.py
@@ -1,1 +1,0 @@
-from .helmcluster import HelmCluster, discover
