--- dask_kubernetes/helm/tests/resources/values.yaml
@@ -1,20 +1,0 @@
-webUI:
-  servicePort: 8087
-
-jupyter:
-  enabled: false
-
-scheduler:
-  serviceType: "ClusterIP"
-  image:
-    repository: "dask-kubernetes"  # Container image repository.
-    tag: "dev"  # Container image tag.
-
-worker:
-  image:
-    repository: "dask-kubernetes"  # Container image repository.
-    tag: "dev"  # Container image tag.
-
-additional_worker_groups:
-  - name: foo
-    replicas: 1
