+++ doc/source/operator_kubecluster.rst
@@ -3,14 +3,10 @@ KubeCluster
 
 .. currentmodule:: dask_kubernetes.operator
 
-.. note::
-
-   As of ``2022.10.0`` the default ``KubeCluster`` class requires the :doc:`Dask Kubernetes Operator <operator>`. For documentation on the classic KubeCluster implementation :doc:`see here <kubecluster>`.
-
 Cluster manager
 ---------------
 
-The operator has a new cluster manager called :class:`dask_kubernetes.operator.KubeCluster` that you can use to conveniently create and manage a Dask cluster in Python. Then connect a Dask :class:`distributed.Client` object to it directly and perform your work.
+The operator has a cluster manager called :class:`dask_kubernetes.operator.KubeCluster` that you can use to conveniently create and manage a Dask cluster in Python. Then connect a Dask :class:`distributed.Client` object to it directly and perform your work.
 
 The goal of the cluster manager is to abstract away the complexity of the Kubernetes resources and provide a clean and simple Python API to manager clusters while still getting all the benefits of the operator.
